Run as non-root in container
Run as non-root in container.
When building the container, set the user ID and group ID of the user in
the container to the user ID and group ID of the user on the host to
share the `logs` directory from the host with the container as a volume.

Simplify analysis script.